+++ autotransform/config/default.py
@@ -7,9 +7,11 @@
 
 # @black_format
 
-"""A ConfigFetcher that uses the autotransform/config/config.ini file to supply configuration."""
+"""A ConfigFetcher that uses config files for configuration."""
 
+import os
 import pathlib
+import subprocess
 from configparser import ConfigParser
 from typing import List, Optional
 
@@ -17,8 +19,13 @@ from autotransform.config.fetcher import ConfigFetcher
 
 
 class DefaultConfigFetcher(ConfigFetcher):
-    """The default configuration fetcher that pulls from the config file.
-    See the sample config in /autotransform/config/sample_config.ini.
+    """The default configuration fetcher that pulls from the config files. Three possible
+    files are used:
+     - a file within the AutoTransform package itself, located at autotransform/config/config.ini
+     - a file within the repo, located at {repo_root}/autotransform/config.ini
+     - a file relative to the current working directory, located at {cwd}/autotransform/config.ini
+    Settings from files later in the list will override settings from files earlier in the list. CWD
+    configs will have the greatest priority, followed by repo configs, then the overall config.
 
     Attributes:
         config (ConfigParser): The parser created from the config file.
@@ -29,25 +36,58 @@ class DefaultConfigFetcher(ConfigFetcher):
     config: ConfigParser
 
     def __init__(self):
-        """A simple constructor that parses the default config file."""
+        """A simple constructor that parses the default config files."""
 
         config = ConfigParser()
-        config.read(self.get_config_path())
+        config_paths = [
+            self.get_package_config_dir(),
+            self.get_repo_config_dir(),
+            self.get_cwd_config_dir(),
+        ]
+        config_paths = [f"{path}/{self.CONFIG_NAME}" for path in config_paths if path is not None]
+        config.read(config_paths)
         self.config = config
 
     @staticmethod
-    def get_config_path() -> str:
-        """Gets the path where the config file is located.
+    def get_repo_config_dir() -> Optional[str]:
+        """Gets the directory where a repo-specific AutoTransform config file would be located. None
+        if git fails. The environment variable AUTO_TRANSFORM_REPO_CONFIG_PATH can be used to set
+        the relative path to the config.ini file for repo configs.
+
+        Returns:
+            Optional[str]: The directory of the repo config file. None if git fails.
+        """
+
+        try:
+            dir_cmd = ["git", "rev-parse", "--show-toplevel"]
+            repo_dir = subprocess.check_output(dir_cmd, encoding="UTF-8").replace("\\", "/").strip()
+            relative_path = os.getenv("AUTO_TRANSFORM_REPO_CONFIG_PATH", "autotransform")
+            return f"{repo_dir}/{relative_path}"
+        except Exception:  # pylint: disable=broad-except
+            return None
+
+    @staticmethod
+    def get_cwd_config_dir() -> str:
+        """Gets the directory where an AutoTransform config file would be located relative to the
+        current working directory. The environment variable AUTO_TRANSFORM_CWD_CONFIG_PATH can be
+        used to set the relative path to the config.ini file for CWD configs.
+
+        Returns:
+            str: The directory of the cwd config file.
+        """
+
+        relative_path = os.getenv("AUTO_TRANSFORM_CWD_CONFIG_PATH", "autotransform")
+        return f"{os.getcwd()}/{relative_path}"
+
+    @staticmethod
+    def get_package_config_dir() -> str:
+        """Gets the directory where a config for the overall AutoTransform package would be located.
 
         Returns:
             str: The path to the config file.
         """
 
-        return (
-            str(pathlib.Path(__file__).parent.resolve()).replace("\\", "/")
-            + "/"
-            + DefaultConfigFetcher.CONFIG_NAME
-        )
+        return str(pathlib.Path(__file__).parent.resolve()).replace("\\", "/")
 
     def get_credentials_github_token(self) -> Optional[str]:
         """Pulls the github authentication token from the config file.